pull game status up into AbstractGame, add ability to abort a game (reset it to a waiting for players state), add a GUID to the game status
